extract matsim from base conf in production-sfbay (#1116)
* add misc summary stats (#1085)

* add misc summary stats

* add null check and change return syntax

* refactoring

* bug fix: wrong field used

* refactor: reuse code

* add totalHoursOfVehicleTrafficDelay in summary stats.

* add summary stats for averagePTWaitingTime

* add bus crowding in summary stats

* reset bus crowding

* add averageTripExpenditure in summary stats

* fixed cost aggregation

* fixed PTCrowding (bus)

* renamed to totalNetCost

* fixed warnings

* reset ptWaiting against personId

* fixed average and update labels

* hoursTraveled should be renamed to travelDurationInSec, as this is not in hours

* remove / 3600 for busCrowding

* fixed hard coding of modes

* fixed hard coding of modes

* rename _person

* fixed warnings and format

* create summary stats for average vehicle delay for motorized leg. (#1095)

* fixed numOfTimesBusTaken in busCrowding

* changed buses list to set

* fixed averageVehicleDelay

* fixed delay stat variable name

* fixed delay stats calculation

* fix n/0 issue

* fix totalVehicles for averageDelays

* fix failing run for sf-light scenario  (#1097)

* fix sf-light run

* fix comment

* can give too much logging when vehicles misconfigured so make exponential

* Cs/#1099 rhm batch and repos (#1104)

* initial attempt to enable co-execution of these processes, still buggy

* debugging

* found bug related to new requests coming in while buffered requests being processed, more robust now

* fixing regression

* another glitch solved

* finally getting simultaneous timeouts working

* fmt

* allow pass when not in batch mode

* ignore failing test

* debug statements mistakently lefts as info

* summary stats modification (#1110)

* Merge remote-tracking branch 'remotes/origin/master' into rajnikantsh/#1008-extract-matsim-conf-4ci

# Conflicts:
#	test/input/sf-light/sf-light-1k.conf
add averageVehicleTravelPerPassengerTrip and averageTravelExpenditure for each activity

* changed variable name

* Cs/#1082 refine analysis (#1109)

* turns into case objects, change delay metric bins

* change delay metric plot  (#1083)

* change plot for delay metric

* Scala fmt

* resolved average delay graph issue

* more descriptive titles

* space

* debugging non-deterministic stuck issue

* fmt

* repairing the fallback BeamPath used for transit trips that can't be … (#1106)

* repairing the fallback BeamPath used for transit trips that can't be routed to avoid exceptions later on

(cherry picked from commit 2cd4bdc)

* scalafmt

* ignore flaky test.

* Cs/#1100 Refactor (#1113)

* decompose some code from RHM into new class to clear things a bit

* fmt

* clean

* extract matsim properties from beam.conf to matsim.conf (#1061)

* extract matsim properties from beam.conf to matsim.conf

* fmt

* set param in beam config and referred in matsim config

* renaming and moving

* better default for iters

* remaining changes in extracting matsim config

* Merge remote-tracking branch 'remotes/origin/master' into rajnikantsh/#1008-extract-matsim-conf-4ci

# Conflicts:
#	test/input/sf-light/sf-light-1k.conf
reset agent number in sf-light-2.5k

* renaming and reogranizing params

* renaming and repositioning of sf-light

* fmt

* extract matsim from base conf

* Merge branch 'master' into rajnikantsh/1115-extract-baseconf-4ci

+++ src/main/scala/beam/agentsim/agents/ridehail/RepositionVehicleRequest.scala
@@ -7,7 +7,7 @@ import org.matsim.vehicles.Vehicle
 
 case class RepositionVehicleRequest(
   passengerSchedule: PassengerSchedule,
-  tick: Double,
+  tick: Int,
   vehicleId: Id[Vehicle],
   rideHailAgent: ActorRef
 )
